Moved eco commands to EcoCMD
